+++ bld/wv/win/c/ntinit.c
@@ -128,7 +128,7 @@ void RestoreHandlers()
 {
 }
 
-long _fork( char *cmd, unsigned len )
+long _fork( char *cmd, size_t len )
 {
     //NYI: check out WinExec
     cmd = cmd; len = len;
